adds xlink connection tracking and menu item to allow dis/reconnections
If a disconnection is detected then no further pings to the xlink server
are sent until the user explicitly reconnects. This keeps the gui responsive
because otherwise each edit would cause a ping and thus a 250ms lag.

A warning is issued if an explicit (re)connection to the remote C64 fails.

If the xlink library is not present on the users system, the menu item is
disabled.